Python basic cert support. AES-CTR mode support.

+++ lib/atcacert/atcacert_der.c
@@ -5,13 +5,13 @@
  * \copyright (c) 2015-2018 Microchip Technology Inc. and its subsidiaries.
  *
  * \page License
- * 
+ *
  * Subject to your compliance with these terms, you may use Microchip software
  * and any derivatives exclusively with Microchip products. It is your
  * responsibility to comply with third party license terms applicable to your
  * use of third party software (including open source software) that may
  * accompany Microchip software.
- * 
+ *
  * THIS SOFTWARE IS SUPPLIED BY MICROCHIP "AS IS". NO WARRANTIES, WHETHER
  * EXPRESS, IMPLIED OR STATUTORY, APPLY TO THIS SOFTWARE, INCLUDING ANY IMPLIED
  * WARRANTIES OF NON-INFRINGEMENT, MERCHANTABILITY, AND FITNESS FOR A
@@ -215,7 +215,7 @@ int atcacert_der_enc_integer(const uint8_t* int_data,
         pad = 1;
     }
 
-    int ret = atcacert_der_enc_length(int_data_size + pad - trim, der_length, &der_length_size);
+    int ret = atcacert_der_enc_length((uint32_t)(int_data_size + pad - trim), der_length, &der_length_size);
     if (ret != ATCACERT_E_SUCCESS)
     {
         return ret;
